==> code/1_processing/1_cleaning.Rmd <==
---
title: "Raw data cleaning"
author: "Leah Lariscy"
output: html_document
---

# Info

This script does the following cleaning steps:

-   Loads raw data files

-   Converts qPCR non-detects to NAs

-   Calculates LOD and LOQ values for all four assays

-   Binds all qPCR data sets with WWTP data and calculates Viral Load

-   Binds all DPH COVID data

# Load packages

```{r,echo=FALSE}
knitr::opts_chunk$set(message=F)
```

```{r,message=FALSE,echo=FALSE}
library(tidyverse)
library(skimr)
library(here)
library(ggplot2)
library(readr)
library(dplyr)
library(stats)
library(ggpmisc)
```

# Load raw data

```{r}
# Load N1 data
n1_stepone_v2 <- read_csv(
  here("data/raw_data/updated_data/stepone_n1_FINAL_UPDATE.csv")) 
    #year 1 data
n1_cfx_v2 <- read_csv(
  here("data/raw_data/updated_data/cfx_n1_FINAL_UPDATE.csv")) 
    #year 2 data
n1_stepone_std <- read_csv(
  here("data/raw_data/standard_curves_stepone.csv")) %>% 
  filter(target == "N1")
    #year 1 N1 standard curve
n1_cfx_std <- read_csv(
  here("data/raw_data/standard_curves_CFX.csv")) %>% 
  filter(target == "N1" & curve == "N1_2")
    #year 2 N1 standard curve

# Load N2 data
n2_stepone_v2 <- read_csv(
  here("data/raw_data/updated_data/stepone_n2_FINAL_UPDATE.csv")) 
    #year 1 data
n2_cfx_v2 <- read_csv(
  here("data/raw_data/updated_data/cfx_n2_FINAL_UPDATE.csv")) 
    #year 2 data
n2_stepone_std <- read_csv(
  here("data/raw_data/standard_curves_stepone.csv")) %>% 
  filter(target == "N2")
    #year 1 N2 standard curve
n2_cfx_std <- read_csv(
  here("data/raw_data/standard_curves_CFX.csv")) %>% 
  filter(target == "N2" & curve == "N2_2")
    #year 2 N1 standard curve

# Load Plant data
plant_v2 <- read_csv(here("data/raw_data/updated_data/plant_data_UPDATED.csv"))

# Load COVID-19 Symptom data
covid_symptom <- read_csv(here("data/raw_data/ga_covid_data/epicurve_symptom_date.csv")) %>% 
  filter(county=="Clarke") %>% 
  select(symptom.date=`symptom date`, 
         cases, moving_avg_cases)

#Load COVID-19 Confirmed Case Data
covid_case <- read_csv(here("data/raw_data/ga_covid_data/epicurve_rpt_date.csv")) %>% 
  filter(county=="Clarke") %>% 
  select(report_date, 
         cases, 
         moving_avg_cases)

#Load COVID-19 Testing Data
covid_testing <- read_csv(here("data/raw_data/ga_covid_data/pcr_antigen_col.csv")) %>% 
  filter(county=="Clarke") %>% 
  select(collection_date = collection_dt, 
         pcr_tests = `ALL PCR tests performed`, 
         pcr_pos = `All PCR positive tests`, 
         pcr_pos_7dma = `7 day percent positive`,
         pcr_pos_14dma = `14 day percent positive`)
```

# Cleaning and merging

## 1. qPCR data

### Convert non-detects to NAs

```{r}
#Select date, collection number, sample id/bio rep, target, and ct
#Convert Undetermined Cts to NAs

#StepOne N1
n1_stepone_clean <- n1_stepone_v2 %>% select(c(sample_date, collection_num, sample_id, target, ct)) %>% 
  mutate(ct=as.numeric(ifelse(ct=="Undetermined", NA, ct)))

#StepOne N2
n2_stepone_clean <- n2_stepone_v2 %>% select(c(sample_date, collection_num, sample_id, target, ct)) %>% 
  mutate(ct=as.numeric(ifelse(ct=="Undetermined", NA, ct)))

#CFX N1
n1_cfx_clean <- n1_cfx_v2 %>% select(c(sample_date, collection_num, sample_id, target, ct)) %>% 
  mutate(ct=as.numeric(ifelse(ct=="Undetermined", NA, ct)))

#CFX N2
n2_cfx_clean <- n2_cfx_v2 %>% select(c(sample_date, collection_num, sample_id, target, ct)) %>% 
  mutate(ct=as.numeric(ifelse(ct=="Undetermined", NA, ct)))
```

### Bind qPCR NA data sets

```{r}
#Bind qpcr data, this will be used for determining detection frequency
qpcr_na <- bind_rows(n1_stepone_clean, n2_stepone_clean, n1_cfx_clean, n2_cfx_clean) %>% 
  mutate(
    date = sample_date,
    facility=substr(sample_id, 1,2), 
    #first two letters in sample_id is treatment facility ID
    biological_replicate=substr(sample_id, nchar(sample_id), nchar(sample_id))) %>% #last number in sample_id is the biological rep
 arrange(date, facility, target, biological_replicate) %>% 
  select(date, facility, target, biological_replicate, collection_num, ct)#select necessary variables
  
#Save to processed data folder
saveRDS(qpcr_na, here("data/processed_data/qpcr_na.rds"))
```

### Calculate LOD and LOQ for qPCR data

```{r}
#Determine the LOD and LOQ by plotting the Normal QQ-Plot
#Code generated by Cody Dailey for Lott et al. 2023
qqnorm.ct.n1.stepone <- qqnorm(n1_stepone_clean$ct, plot.it = T) %>% as.data.frame()
qqnorm.ct.n2.stepone <- qqnorm(n2_stepone_clean$ct, plot.it = T) %>% as.data.frame()
qqnorm.ct.n1.cfx <- qqnorm(n1_cfx_clean$ct, plot.it = T) %>% as.data.frame()
qqnorm.ct.n2.cfx <- qqnorm(n2_cfx_clean$ct, plot.it = T) %>% as.data.frame()

#Create function to compute LOD and lOQ based on plots
qqnorm.Explorer.ct <- function(qqnorm.ct){
  qqnorm.ct <- qqnorm.ct[which(complete.cases(qqnorm.ct)),]
  qqnorm.ct <- qqnorm.ct[order(qqnorm.ct$x),]
  qqnorm.ct <- cbind(qqnorm.ct, rbind(NA, qqnorm.ct[-nrow(qqnorm.ct),])) %>% setNames(., nm = c("x", "y", "x-1", "y-1"))
  qqnorm.ct %<>% mutate(rise = y-`y-1`, run = x-`x-1`) %>% mutate(slope = rise / run)
  qqnorm.ct$lod <- NA
  qqnorm.ct$loq <- NA
  prev.slope <- 1
  lod.found <- 0
  for(i in nrow(qqnorm.ct):2){
    if(lod.found==0){
      if(qqnorm.ct$slope[i]<1 & prev.slope <1){
        qqnorm.ct$lod[i] <- 1
        lod.found <- 1
      }else{
        prev.slope <- qqnorm.ct$slope[i]
      }
    }
    if(lod.found==1){
      if(qqnorm.ct$slope[i]>1){
        qqnorm.ct$loq[i] <- 1
        break
      }else{
        prev.slope <- qqnorm.ct$slope[i]
      }
    }
  }
  lod.ct <- qqnorm.ct$y[which(qqnorm.ct$lod==1)]
  loq.ct <- qqnorm.ct$y[which(qqnorm.ct$loq==1)]
  return(list(qqnorm.dataset = qqnorm.ct, lod = lod.ct, loq = loq.ct))
}

#Run function on each data set
qqnorm.ct.n1.stepone <- qqnorm.Explorer.ct(qqnorm.ct.n1.stepone)
qqnorm.ct.n2.stepone <- qqnorm.Explorer.ct(qqnorm.ct.n2.stepone)
qqnorm.ct.n1.cfx <- qqnorm.Explorer.ct(qqnorm.ct.n1.cfx)
qqnorm.ct.n2.cfx <- qqnorm.Explorer.ct(qqnorm.ct.n2.cfx)

#Save LOD and LOQ for each data set
n1_stepone_lod <- qqnorm.ct.n1.stepone$lod
n1_stepone_loq <- qqnorm.ct.n1.stepone$loq
n2_stepone_lod <- qqnorm.ct.n2.stepone$lod
n2_stepone_loq <- qqnorm.ct.n2.stepone$loq

n1_cfx_lod <- qqnorm.ct.n1.cfx$lod
n1_cfx_loq <- qqnorm.ct.n1.cfx$loq
n2_cfx_lod <- qqnorm.ct.n2.cfx$lod
n2_cfx_loq <- qqnorm.ct.n2.cfx$loq

#combine values
limit_cq_values <- list("Y1 N1 LoQ" = n1_stepone_loq,
            "Y1 N2 LoQ" = n2_stepone_loq,
            "Y2 N1 LoQ" = n1_cfx_loq,
            "Y2 N2 LoQ" = n2_cfx_loq,
            "Y1 N1 LoD" = n1_stepone_lod,
            "Y1 N2 LoD" = n2_stepone_lod,
            "Y2 N1 LoD" = n1_cfx_lod,
            "Y2 N2 LoD" = n2_cfx_lod)

limit_cq_values_combined <- enframe(limit_cq_values, name = "Limits", value = "Cq")
saveRDS(limit_cq_values_combined, here("data/processed_data/wastewater/loqs_lods_cq.rds"))
```

### Calculate standard curves

```{r}
#stepone N1
n1_stepone_lm <- lm(ct~log10(quantity), data = n1_stepone_std)
n1_stepone_intercept <- coef(n1_stepone_lm)[1]
n1_stepone_slope <- coef(n1_stepone_lm)[2]

print(paste("Cq =", round(n1_stepone_slope,2), "x", "log10quant", "+", round(n1_stepone_intercept,2)))

#stepone N2
n2_stepone_lm <- lm(ct~log10(quantity), data = n2_stepone_std)
n2_stepone_intercept <- coef(n2_stepone_lm)[1]
n2_stepone_slope <- coef(n2_stepone_lm)[2]

print(paste("Cq =", round(n2_stepone_slope,2), "x", "log10quant", "+", round(n2_stepone_intercept,2)))

#CFX N1
n1_cfx_lm <- lm(ct~log10(quantity), data = n1_cfx_std)
n1_cfx_intercept <- coef(n1_cfx_lm)[1]
n1_cfx_slope <- coef(n1_cfx_lm)[2]

print(paste("Cq =", round(n1_cfx_slope,2), "x", "log10quant", "+", round(n1_cfx_intercept,2)))

#CFX N2
n2_cfx_lm <- lm(ct~log10(quantity), data = n2_cfx_std)
n2_cfx_intercept <- coef(n2_cfx_lm)[1]
n2_cfx_slope <- coef(n2_cfx_lm)[2]

print(paste("Cq =", round(n2_cfx_slope,2), "x", "log10quant", "+", round(n2_cfx_intercept,2)))
```

```{r}
n1_y1_std_plot <- n1_stepone_std %>%
  ggplot(aes(log10(quantity), ct)) +
  geom_point() +
  geom_smooth(method = "lm",alpha = 0) +
  ggthemes::theme_clean() +
  stat_poly_eq(use_label("eq","R2")) +
  ggtitle("N1 Year 1")
ggsave(here("figures/std_curves/n1_y1.png"))

n2_y1_std_plot <- n2_stepone_std %>% 
  ggplot(aes(log10(quantity), ct)) +
  geom_point() +
  geom_smooth(method = "lm",alpha = 0) +
  ggthemes::theme_clean() +
  stat_poly_eq(use_label("eq","R2")) +
  ggtitle("N2 Year 1")
ggsave(here("figures/std_curves/n2_y1.png"))

n1_y2_std_plot <- n1_cfx_std %>% 
  ggplot(aes(log10(quantity), ct)) +
  geom_point() +
  geom_smooth(method = "lm",alpha = 0) +
  ggthemes::theme_clean() +
  stat_poly_eq(use_label("eq","R2")) +
  ggtitle("N1 Year 2+")
ggsave(here("figures/std_curves/n1_y2.png"))

n2_y2_std_plot <- n2_cfx_std %>% 
  ggplot(aes(log10(quantity), ct)) +
  geom_point() +
  geom_smooth(method = "lm",alpha = 0) +
  ggthemes::theme_clean() +
  stat_poly_eq(use_label("eq","R2")) +
  ggtitle("N2 Year 2+")
ggsave(here("figures/std_curves/n2_y2.png"))

ggpubr::ggarrange(n1_y1_std_plot,
                  n1_y2_std_plot,
                  n2_y1_std_plot,
                  n2_y2_std_plot,
                  nrow = 2,
                  ncol = 2, 
                  labels = "auto")
ggsave(here("figures/std_curves/all.tiff"),
       height = 7,
       width = 9)
```

### Calculate half of cp/uL LoQ

```{r}
#Calculate half LOQ for each assay, to assign to rows greater than LoQ
#Units = copies per uL of reaction (multiply by 10 for starting quant)

n1_stepone_half_loq <- as.numeric(
  10^((n1_stepone_loq-n1_stepone_intercept)/n1_stepone_slope)/2)

n2_stepone_half_loq <- as.numeric(
  10^((n2_stepone_loq-n2_stepone_intercept)/n2_stepone_slope)/2)

n1_cfx_half_loq <- as.numeric(
  10^((n1_cfx_loq-n1_cfx_intercept)/n1_cfx_slope)/2)

n2_cfx_half_loq <- as.numeric(
  10^((n2_cfx_loq-n2_cfx_intercept)/n2_cfx_slope)/2)
```

### Calculate half of cp/uL LoD

```{r}
n1_stepone_half_lod <- as.numeric(
  10^((n1_stepone_lod-n1_stepone_intercept)/n1_stepone_slope)/2)

n2_stepone_half_lod <- as.numeric(
  10^((n2_stepone_lod-n2_stepone_intercept)/n2_stepone_slope)/2)

n1_cfx_half_lod <- as.numeric(
  10^((n1_cfx_lod-n1_cfx_intercept)/n1_cfx_slope)/2)

n2_cfx_half_lod <- as.numeric(
  10^((n2_cfx_lod-n2_cfx_intercept)/n2_cfx_slope)/2)
```

```{r}
#limits in total copies per rxn
n1_stepone_lod_rxn <- as.numeric(
  10^((n1_stepone_lod-n1_stepone_intercept)/n1_stepone_slope)*20)

n2_stepone_lod_rxn <- as.numeric(
  10^((n2_stepone_lod-n2_stepone_intercept)/n2_stepone_slope)*20)

n1_cfx_lod_rxn <- as.numeric(
  10^((n1_cfx_lod-n1_stepone_intercept)/n1_cfx_slope)*20)

n2_cfx_lod_rxn <- as.numeric(
  10^((n2_cfx_lod-n2_stepone_intercept)/n2_cfx_slope)*20)
```

### Calculate LoD and LoQ in copies/L

```{r}
#loq
n1_stepone_loq_liters <- as.numeric(
  10^((n1_stepone_loq-n1_stepone_intercept)/n1_stepone_slope)
  *20/2*25/3*60/280*1000*1000)
n1_stepone_loq_liters %>% format(scientific = T)

n2_stepone_loq_liters <- as.numeric(
  10^((n2_stepone_loq-n2_stepone_intercept)/n2_stepone_slope)*20/2*25/3*60/280*1000*1000)
n2_stepone_loq_liters %>% format(scientific = T)

n1_cfx_loq_liters <- as.numeric(
  10^((n1_cfx_loq-n1_cfx_intercept)/n1_cfx_slope)*20/5*60/280*1000*1000)
n1_cfx_loq_liters %>% format(scientific = T)

n2_cfx_loq_liters <- as.numeric(
  10^((n2_cfx_loq-n2_cfx_intercept)/n2_cfx_slope)*20/5*60/280*1000*1000)
n2_cfx_loq_liters %>% format(scientific = T)

#lod
n1_stepone_lod_liters <- as.numeric(
  10^((n1_stepone_lod-n1_stepone_intercept)/n1_stepone_slope)*20/2*25/3*60/280*1000*1000)
n1_stepone_lod_liters %>% format(scientific = T)

n2_stepone_lod_liters <- as.numeric(
  10^((n2_stepone_lod-n2_stepone_intercept)/n2_stepone_slope)*20/2*25/3*60/280*1000*1000)
n2_stepone_lod_liters %>% format(scientific = T)

n1_cfx_lod_liters <- as.numeric(
  10^((n1_cfx_lod-n1_cfx_intercept)/n1_cfx_slope)*20/5*60/280*1000*1000)
n1_cfx_lod_liters %>% format(scientific = T)

n2_cfx_lod_liters <- as.numeric(
  10^((n2_cfx_lod-n2_cfx_intercept)/n2_cfx_slope)*20/5*60/280*1000*1000)
n2_cfx_lod_liters %>% format(scientific = T)

#combine values
limit_values <- list("Y1 N1 LoQ" = n1_stepone_loq_liters,
            "Y1 N2 LoQ" = n2_stepone_loq_liters,
            "Y2 N1 LoQ" = n1_cfx_loq_liters,
            "Y2 N2 LoQ" = n2_cfx_loq_liters,
            "Y1 N1 LoD" = n1_stepone_lod_liters,
            "Y1 N2 LoD" = n2_stepone_lod_liters,
            "Y2 N1 LoD" = n1_cfx_lod_liters,
            "Y2 N2 LoD" = n2_cfx_lod_liters)

limit_values_combined <- enframe(limit_values, name = "Limits", value = "Copies/L")
saveRDS(limit_values_combined, here("data/processed_data/wastewater/loqs_lods_copies.rds"))
```

### Transform Cq NAs, calculate copies/uL

```{r}
#Use standard curve slope for each target to calculate copies per uL per rxn
#Replace cp/uL for observations with Cqs > LoQ and < LoD with half cp/uL LoQ
#Replace cp/uL for observations with Cqs > LoD with half cp/uL LoD

#StepOne N1 ################################################################
n1_stepone_ct <- n1_stepone_clean %>% 
  mutate(nondetect = ifelse(is.na(ct), "nondetect", "detect"),
         ct_loq = ifelse(ct < n1_stepone_loq, "quantifiable", "nonquant"),
         ct_lod = ifelse(
           ct < n1_stepone_lod, "detectable", "nondetectable"))

#transform nondetectable rxns
n1_stepone_nondetect <- n1_stepone_ct %>% 
  filter(nondetect == "nondetect" | ct_lod == "nondetectable") %>% 
  mutate(copy_num_uL_rxn = n1_cfx_half_lod,
         copy_num_L = copy_num_uL_rxn*20/2*25/3*60/280*1000*1000) 

#transform nonquant but detectable rxns
n1_stepone_nonquant <- n1_stepone_ct %>% 
  filter(ct_loq == "nonquant" & ct_lod == "detectable") %>% 
  mutate(copy_num_uL_rxn = n1_cfx_half_loq,
         copy_num_L = copy_num_uL_rxn*20/2*25/3*60/280*1000*1000) 

#transform quantifiable rxns
n1_stepone_quant <- n1_stepone_ct %>% 
  filter(ct_loq == "quantifiable" & ct_lod == "detectable") %>% 
  mutate(copy_num_uL_rxn = 10^((ct-n1_stepone_intercept)/n1_stepone_slope),
         copy_num_L = copy_num_uL_rxn*20/2*25/3*60/280*1000*1000)

#bind rows
n1_stepone_1 <- bind_rows(n1_stepone_nondetect,
                          n1_stepone_nonquant,
                          n1_stepone_quant) %>% 
                arrange(sample_date)

#StepOne N2 ################################################################
n2_stepone_ct <- n2_stepone_clean %>% 
  mutate(nondetect = ifelse(is.na(ct), "nondetect", "detect"),
         ct_loq = ifelse(ct < n2_stepone_loq, "quantifiable", "nonquant"),
         ct_lod = ifelse(
           ct < n2_stepone_lod, "detectable", "nondetectable"))

#transform nondetectable rxns
n2_stepone_nondetect <- n2_stepone_ct %>% 
  filter(nondetect == "nondetect" | ct_lod == "nondetectable") %>% 
  mutate(copy_num_uL_rxn = n2_cfx_half_lod,
         copy_num_L = copy_num_uL_rxn*20/2*25/3*60/280*1000*1000) 

#transform nonquant but detectable rxns
n2_stepone_nonquant <- n2_stepone_ct %>% 
  filter(ct_loq == "nonquant" & ct_lod == "detectable") %>% 
  mutate(copy_num_uL_rxn = n2_cfx_half_loq,
         copy_num_L = copy_num_uL_rxn*20/2*25/3*60/280*1000*1000) 

#transform quantifiable rxns
n2_stepone_quant <- n2_stepone_ct %>% 
  filter(ct_loq == "quantifiable" & ct_lod == "detectable") %>% 
  mutate(copy_num_uL_rxn = 10^((ct-n2_stepone_intercept)/n2_stepone_slope),
         copy_num_L = copy_num_uL_rxn*20/2*25/3*60/280*1000*1000)

#bind rows
n2_stepone_1 <- bind_rows(n2_stepone_nondetect,
                          n2_stepone_nonquant,
                          n2_stepone_quant) %>% 
                arrange(sample_date)

#CFX N1 ####################################################################
n1_cfx_ct <- n1_cfx_clean %>% 
  mutate(nondetect = ifelse(is.na(ct), "nondetect", "detect"),
         ct_loq = ifelse(ct < n1_cfx_loq, "quantifiable", "nonquant"),
         ct_lod = ifelse(
           ct < n1_cfx_lod, "detectable", "nondetectable"))

#transform nondetectable rxns
n1_cfx_nondetect <- n1_cfx_ct %>% 
  filter(nondetect == "nondetect" | ct_lod == "nondetectable") %>% 
  mutate(copy_num_uL_rxn = n1_cfx_half_lod,
         copy_num_L = copy_num_uL_rxn*20/5*60/280*1000*1000) 

#transform nonquant but detectable rxns
n1_cfx_nonquant <- n1_cfx_ct %>% 
  filter(ct_loq == "nonquant" & ct_lod == "detectable") %>% 
  mutate(copy_num_uL_rxn = n1_cfx_half_loq,
         copy_num_L = copy_num_uL_rxn*20/5*60/280*1000*1000) 

#transform quantifiable rxns
n1_cfx_quant <- n1_cfx_ct %>% 
  filter(ct_loq == "quantifiable" & ct_lod == "detectable") %>% 
  mutate(copy_num_uL_rxn = 10^((ct-n1_cfx_intercept)/n1_cfx_slope),
         copy_num_L = copy_num_uL_rxn*20/5*60/280*1000*1000)

#bind rows
n1_cfx_1 <- bind_rows(n1_cfx_nondetect,
                          n1_cfx_nonquant,
                          n1_cfx_quant) %>% 
                arrange(sample_date) 

#CFX N2 ####################################################################
n2_cfx_ct <- n2_cfx_clean %>% 
  mutate(nondetect = ifelse(is.na(ct), "nondetect", "detect"),
         ct_loq = ifelse(ct < n2_cfx_loq, "quantifiable", "nonquant"),
         ct_lod = ifelse(
           ct < n2_cfx_lod, "detectable", "nondetectable"))

#transform nondetectable rxns
n2_cfx_nondetect <- n2_cfx_ct %>% 
  filter(nondetect == "nondetect" | ct_lod == "nondetectable") %>% 
  mutate(copy_num_uL_rxn = n2_cfx_half_lod,
         copy_num_L = copy_num_uL_rxn*20/5*60/280*1000*1000) 

#transform nonquant but detectable rxns
n2_cfx_nonquant <- n2_cfx_ct %>% 
  filter(ct_loq == "nonquant" & ct_lod == "detectable") %>% 
  mutate(copy_num_uL_rxn = n2_cfx_half_loq,
         copy_num_L = copy_num_uL_rxn*20/5*60/280*1000*1000) 

#transform quantifiable rxns
n2_cfx_quant <- n2_cfx_ct %>% 
  filter(ct_loq == "quantifiable" & ct_lod == "detectable") %>% 
  mutate(copy_num_uL_rxn = 10^((ct-n2_cfx_intercept)/n2_cfx_slope),
         copy_num_L = copy_num_uL_rxn*20/5*60/280*1000*1000)

#bind rows
n2_cfx_1 <- bind_rows(n2_cfx_nondetect,
                          n2_cfx_nonquant,
                          n2_cfx_quant) %>% 
                arrange(sample_date) 

saveRDS(n1_stepone_1, here("data/processed_data/wastewater/n1_stepone_unsummarized.rds"))
saveRDS(n2_stepone_1, here("data/processed_data/wastewater/n2_stepone_unsummarized.rds"))
saveRDS(n1_cfx_1, here("data/processed_data/wastewater/n1_cfx_unsummarized.rds"))
saveRDS(n1_cfx_1, here("data/processed_data/wastewater/n1_cfx_unsummarized.rds"))

```

```{r}
#Visualize distributions
#N1 StepOne
n1_stepone_1 %>% filter(ct<n1_stepone_lod) %>% ggplot(aes(ct)) + geom_histogram()

#N2 StepOne
n2_stepone_1 %>% filter(ct<n2_stepone_lod) %>% ggplot(aes(ct)) + geom_histogram()

#N1 CFX
n1_cfx_1 %>% filter(ct<n1_cfx_lod) %>% ggplot(aes(ct)) + geom_histogram()

#N2 CFX
n2_cfx_1 %>% filter(ct<n2_cfx_lod) %>% ggplot(aes(ct)) + geom_histogram()

#year one data has many non-detects, makes sense for earlier in pandemic
```

### Bind qPCR data sets

```{r}
#Bind qpcr data
#this will be used for calculating viral load
qpcr_all <- bind_rows(n1_stepone_1, n2_stepone_1, n1_cfx_1, n2_cfx_1) %>% 
  mutate(
    date = sample_date,
    facility=substr(sample_id, 1,2), #first two letters in sample_id is facility ID
    biological_replicate=substr(sample_id, nchar(sample_id), nchar(sample_id))) %>% 
  #last number in sample_id is the biological rep
 arrange(date, facility, target, biological_replicate) %>% 
  select(!sample_id)#select necessary variables
  
#Save to processed data folder
saveRDS(qpcr_all, here("data/processed_data/qpcr_all1.rds"))
qpcr_all
```

## 2. Combine plant and qPCR data sets

```{r}
#Change plant variable names to match qPCR names, remove notes variable, convert millions of gallons to liters
plant_v2 <- plant_v2 %>% 
  rename(collection_num = Collection, facility = wrf) %>% 
  mutate(influent_flow_L = influent_flow_mg*1e6*231*(0.0254^3)*1000) 

#Merge and mutate
qpcr_plant_all <- merge(qpcr_all, plant_v2, 
                        by = c("collection_num", "facility", "date"),
                        all = T) %>% 
  mutate(facility = as.factor(facility), #code each facility as a factor
         facility = recode(facility, NO = "A", MI = "B", CC = "C"), 
         #de-identify treatment facility
         facility = ordered(facility, levels = c("A", "B", "C")), 
         #set facility factor levels
         viral_load = copy_num_L*influent_flow_L) %>% 
        #calculate total copies per day (viral load) using flow data
         drop_na(target) %>% 
         select(!c(notes))

qpcr_plant_all

#Save to processed data folder
saveRDS(qpcr_plant_all, here("data/processed_data/qpcr_plant_all.rds"))
```

## 3. Combine DPH COVID data sets

```{r}
covid <- full_join(
  covid_symptom%>%
    select(cases.symptom.onset=cases, date=symptom.date), 
  covid_case%>%
    select(cases.reported=cases, date=report_date), 
  by = "date"
) %>% 
  full_join(
    covid_testing%>%
      rename(date=collection_date), 
    by="date"
  ) %>%
  select(date, cases.symptom.onset, cases.reported, pcr_tests, pcr_pos, pcr_pos_7dma, pcr_pos_14dma) %>% 

  mutate(prop_pos = pcr_pos/pcr_tests)

#Save to processed data folder
saveRDS(covid, here("data/processed_data/all_covid_combined.rds"))
covid
```
