==> code/1_processing/2_preprocessing.Rmd <==
---
title: "Preprocessing"
author: "Leah Lariscy"
output: html_document
---

## Info

This pre-processing script does the following:

-   Viral load technical reps are averaged

-   Detection frequency technical reps are summed

-   Biological reps are randomly subset at n = 6,5,4,3,2,1 for VL and DF

-   Viral load is averaged per WRF, target, week

-   Detection frequency is calculated per WRF, target, week

-   Data is widened so each WRF/gene target combo is a unique variable

-   COVID data is summed by week

-   All WBE and COVID data are combined

-   Detection frequency ratio is calculated

**Data is now ready for further analysis**

## Load packages

```{r}
knitr::opts_chunk$set(message=F)
```

```{r}
library(tidyverse)
library(here)
library(tsibble)
library(skimr)
library(ggpubr)
library(DescTools)
```

## Load data

```{r}
#WBE file from processing script
wbe <- read_rds(here("data/processed_data/qpcr_plant_all.rds"))

#COVID file from processing script
covid <- read_rds(here("data/processed_data/all_covid_combined.rds"))

#Hospitalization data from processing script
hospital <- read_rds(here("data/processed_data/athens_hospitalizations.rds"))

#Load qPCR file with ct NAs from processing script
qpcr_na <- read_rds(here("data/processed_data/qpcr_na.rds"))
```

## WBE wrangling

```{r}
# Counting the number of unique collection dates
unique_count <- length(unique(wbe$date))

# Display the result
print(paste("Number of unique observations:", unique_count))
```

### 1. Average VL technical reps

This will be used to subset data by n of biological replicates

```{r}
wbe_vl_avg <- wbe %>% 
  group_by(date,facility,target,biological_replicate) %>% 
  summarize(avg_viral_load = exp(mean(log(viral_load)))) %>% 
  ungroup()

saveRDS(wbe_vl_avg, here("data/processed_data/wastewater/viral_load_tech_rep_avg.rds"))
```

#### Subset bio reps & summarize

```{r}
# n = 6
set.seed(13)
vl_avg_n6 <- wbe_vl_avg %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=6) %>% 
  summarize(avg_viral_load = exp(mean(log(avg_viral_load))))

#n = 5
set.seed(13)
vl_avg_n5 <- wbe_vl_avg %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=5) %>% 
  summarize(avg_viral_load = exp(mean(log(avg_viral_load))))

#n = 4
set.seed(13)
vl_avg_n4 <- wbe_vl_avg %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=4) %>% 
  summarize(avg_viral_load = exp(mean(log(avg_viral_load))))

#n = 3
set.seed(13)
vl_avg_n3 <- wbe_vl_avg %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=3) %>% 
  summarize(avg_viral_load = exp(mean(log(avg_viral_load))))

#n = 2
set.seed(13)
vl_avg_n2 <- wbe_vl_avg %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=2) %>% 
  summarize(avg_viral_load = exp(mean(log(avg_viral_load))))

#n = 1
set.seed(13)
vl_avg_n1 <- wbe_vl_avg %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=1) %>% 
  summarize(avg_viral_load = exp(mean(log(avg_viral_load))))

saveRDS(vl_avg_n6, here("data/processed_data/wastewater/viral_load_daily_avg.rds"))
```

### 2. Sum DF technical reps

```{r}
wbe_df_sum <- qpcr_na %>% 
  mutate(facility = as.factor(facility), #code each facility as a factor
         facility = recode(facility, NO = "A", MI = "B", CC = "C"), 
         #de-identify treatment facility
         facility = ordered(facility, levels = c("A", "B", "C"))) %>% 
  filter(between(date, as.Date("2020-06-30"), as.Date("2023-01-04"))) %>% 
        select(!collection_num) %>% 
        group_by(date,facility,target,biological_replicate) %>% 
        summarize(n = n(), 
                  n_miss = sum(is.na(ct)),
                  n_pos = n-n_miss) %>% 
        ungroup()
```

#### Subset bio reps & summarize by date

```{r}
# n = 6
set.seed(13)
df_sum_n6 <- wbe_df_sum %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=6) %>% 
  summarize(n = sum(n), 
            n_miss = sum(n_miss),
            n_pos = n-n_miss) %>% 
  ungroup() %>% 
  select(!n_miss)

# n = 5
set.seed(13)
df_sum_n5 <- wbe_df_sum %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=5) %>% 
  summarize(n = sum(n), 
            n_miss = sum(n_miss),
            n_pos = n-n_miss) %>% 
  ungroup() %>% 
  select(!n_miss)

# n = 4
set.seed(13)
df_sum_n4 <- wbe_df_sum %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=4) %>% 
  summarize(n = sum(n), 
            n_miss = sum(n_miss),
            n_pos = n-n_miss) %>% 
  ungroup() %>% 
  select(!n_miss)

# n = 3
set.seed(13)
df_sum_n3 <- wbe_df_sum %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=3) %>% 
  summarize(n = sum(n), 
            n_miss = sum(n_miss),
            n_pos = n-n_miss) %>% 
  ungroup() %>% 
  select(!n_miss)

# n = 2
set.seed(13)
df_sum_n2 <- wbe_df_sum %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=2) %>% 
  summarize(n = sum(n), 
            n_miss = sum(n_miss),
            n_pos = n-n_miss) %>% 
  ungroup() %>% 
  select(!n_miss)

# n = 1
set.seed(13)
df_sum_n1 <- wbe_df_sum %>% 
  group_by(date,facility,target) %>% 
  slice_sample(n=1) %>% 
  summarize(n = sum(n), 
            n_miss = sum(n_miss),
            n_pos = n-n_miss) %>% 
  ungroup() %>% 
  select(!n_miss)

saveRDS(df_sum_n6, here("data/processed_data/wastewater/detection_frequency_daily.rds"))
```

### 3. Widen viral load data

Then, average by week

VL data ready to be combined with DF data after this

```{r}
# n = 6
vl_n6_week <- vl_avg_n6 %>% 
  select(c(date,avg_viral_load,facility,target)) %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(avg_viral_load),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise(across(c(2,3,4,5,6,7), ~ exp(mean(log(.)))))

# n = 5
vl_n5_week <- vl_avg_n5 %>% 
  select(c(date,avg_viral_load,facility,target)) %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(avg_viral_load),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise(across(c(2,3,4,5,6,7), ~ exp(mean(log(.)))))

# n = 4
vl_n4_week <- vl_avg_n4 %>% 
  select(c(date,avg_viral_load,facility,target)) %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(avg_viral_load),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise(across(c(2,3,4,5,6,7), ~ exp(mean(log(.)))))

# n = 3
vl_n3_week <- vl_avg_n3 %>% 
  select(c(date,avg_viral_load,facility,target)) %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(avg_viral_load),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise(across(c(2,3,4,5,6,7), ~ exp(mean(log(.)))))

# n = 2
vl_n2_week <- vl_avg_n2 %>% 
  select(c(date,avg_viral_load,facility,target)) %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(avg_viral_load),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise(across(c(2,3,4,5,6,7), ~ exp(mean(log(.)))))

# n = 1
vl_n1_week <- vl_avg_n1 %>% 
  select(c(date,avg_viral_load,facility,target)) %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(avg_viral_load),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise(across(c(2,3,4,5,6,7), ~ exp(mean(log(.)))))

saveRDS(vl_n6_week, here("data/processed_data/wastewater/viral_load_weekly_avg.rds"))
```

### 4. Widen detection frequency data

Then, sum by week

AP data ready to be combined with VL data after this

```{r}
# n = 6
df_n6_week <- df_sum_n6 %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(n, n_pos),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise_at(vars(2:13), sum, ra.rm = F)

# n = 5
df_n5_week <- df_sum_n5 %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(n, n_pos),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise_at(vars(2:13), sum, ra.rm = F)

# n = 4
df_n4_week <- df_sum_n4 %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(n, n_pos),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise_at(vars(2:13), sum, ra.rm = F)

# n = 3
df_n3_week <- df_sum_n3 %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(n, n_pos),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise_at(vars(2:13), sum, ra.rm = F)

# n = 2
df_n2_week <- df_sum_n2 %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(n, n_pos),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise_at(vars(2:13), sum, ra.rm = F)

# n = 1
df_n1_week <- df_sum_n1 %>% 
  pivot_wider(names_from = c(facility,target),
              values_from = c(n, n_pos),
              names_sep = "_") %>% 
  mutate(date = as_date(date), 
         week = yearweek(date, week_start = 7)) %>% 
  group_by(week) %>% 
  summarise_at(vars(2:13), sum, ra.rm = F)
```

### 5. Combine VL and AP data

WBE data ready to be combined with COVID data

```{r}
wbe_n6_week <- left_join(df_n6_week,vl_n6_week, by = "week") #n=6

wbe_n5_week <- left_join(df_n5_week,vl_n5_week, by = "week") #n=5

wbe_n4_week <- left_join(df_n4_week,vl_n4_week, by = "week") #n=4

wbe_n3_week <- left_join(df_n3_week,vl_n3_week, by = "week") #n=3

wbe_n2_week <- left_join(df_n2_week,vl_n2_week, by = "week") #n=2

wbe_n1_week <- left_join(df_n1_week,vl_n1_week, by = "week") #n=1
```

## COVID Wrangling

### 1. Summarize by week

```{r}
covid_daily <- covid %>% 
  filter(between(date, as.Date("2020-06-30"), as.Date("2023-01-04"))) %>%  #filter for surveillance time series
  mutate(date = as_date(date),
         week = yearweek(date, week_start = 7))

covid_weekly <- covid_daily %>% 
  select(!c(cases.symptom.onset,
            pcr_pos_7dma,pcr_pos_14dma,prop_pos,date)) %>% 
  #select only week, pcr test, pcr pos, and cases reported
  group_by(week) %>% 
  summarise_at(vars(1:3),sum,na.rm = F) %>% 
  #sum tests and positives by week
  ungroup() %>% 
  mutate(cases.reported = 
           round((cases.reported*100000)/128711, digits = 2),
         pcr_tests = 
           round((pcr_tests*100000)/128711, digits = 2))
  #convert total reported cases and tests administered to cases per 100,000
  #ACC population = 128,711

covid_weekly_nonnormalized <- covid_daily %>% 
  select(!c(cases.symptom.onset,
            pcr_pos_7dma,pcr_pos_14dma,prop_pos,date)) %>% 
  #select only week, pcr test, pcr pos, and cases reported
  group_by(week) %>% 
  summarise_at(vars(1:3),sum,na.rm = F) %>% 
  #sum tests and positives by week
  ungroup()

#save data for modelling
saveRDS(covid_weekly, 
        here("data/processed_data/clinical/covid_cases_weekly.rds")) 

#save data for summary stats
saveRDS(covid_weekly_nonnormalized, 
        here("data/processed_data/clinical/covid_cases_weekly_nonnorm.rds")) 
```

### 2. Combine WBE and COVID data

```{r}
#this chunk also calculates the detection frequency ratio
# n = 6
wbe_covid_n6_week <- left_join(wbe_n6_week, covid_weekly, by = "week") %>% drop_na(week) %>% 
  mutate_at(vars(14:19),log10) %>% 
  mutate(A_N1_POS = n_pos_A_N1/n_A_N1,
         A_N2_POS = n_pos_A_N2/n_A_N2,
         B_N1_POS = n_pos_B_N1/n_B_N1,
         B_N2_POS = n_pos_B_N2/n_B_N2,
         C_N1_POS = n_pos_C_N1/n_C_N1,
         C_N2_POS = n_pos_C_N2/n_C_N2,
         log10_cases = log10(cases.reported)) %>% 
  mutate(date = as.Date(week)) %>% 
  mutate(qpcr = if_else(date > "2021-06-29","year2","year1"))

saveRDS(wbe_covid_n6_week,
        here("data/processed_data/wbe_covid_n6_week.rds"))
############################################################################
# n = 5
wbe_covid_n5_week <- left_join(wbe_n5_week, covid_weekly, by = "week") %>% drop_na(week) %>% 
  mutate_at(vars(14:19),log10) %>% 
  mutate(A_N1_POS = n_pos_A_N1/n_A_N1,
         A_N2_POS = n_pos_A_N2/n_A_N2,
         B_N1_POS = n_pos_B_N1/n_B_N1,
         B_N2_POS = n_pos_B_N2/n_B_N2,
         C_N1_POS = n_pos_C_N1/n_C_N1,
         C_N2_POS = n_pos_C_N2/n_C_N2,
         log10_cases = log10(cases.reported)) %>% 
  mutate(date = as.Date(week)) %>% 
  mutate(qpcr = if_else(date > "2021-06-29","year2","year1"))

saveRDS(wbe_covid_n5_week,
        here("data/processed_data/wbe_covid_n5_week.rds"))
############################################################################
# n = 4
wbe_covid_n4_week <- left_join(wbe_n4_week, covid_weekly, by = "week") %>% drop_na(week) %>% 
  mutate_at(vars(14:19),log10) %>% 
  mutate(A_N1_POS = n_pos_A_N1/n_A_N1,
         A_N2_POS = n_pos_A_N2/n_A_N2,
         B_N1_POS = n_pos_B_N1/n_B_N1,
         B_N2_POS = n_pos_B_N2/n_B_N2,
         C_N1_POS = n_pos_C_N1/n_C_N1,
         C_N2_POS = n_pos_C_N2/n_C_N2,
         log10_cases = log10(cases.reported)) %>% 
  mutate(date = as.Date(week)) %>% 
  mutate(qpcr = if_else(date > "2021-06-29","year2","year1"))

saveRDS(wbe_covid_n4_week,
        here("data/processed_data/wbe_covid_n4_week.rds"))
############################################################################
# n = 3
wbe_covid_n3_week <- left_join(wbe_n3_week, covid_weekly, by = "week") %>% drop_na(week) %>% 
  mutate_at(vars(14:19),log10) %>% 
  mutate(A_N1_POS = n_pos_A_N1/n_A_N1,
         A_N2_POS = n_pos_A_N2/n_A_N2,
         B_N1_POS = n_pos_B_N1/n_B_N1,
         B_N2_POS = n_pos_B_N2/n_B_N2,
         C_N1_POS = n_pos_C_N1/n_C_N1,
         C_N2_POS = n_pos_C_N2/n_C_N2,
         log10_cases = log10(cases.reported)) %>% 
  mutate(date = as.Date(week)) %>% 
  mutate(qpcr = if_else(date > "2021-06-29","year2","year1"))

saveRDS(wbe_covid_n3_week,
        here("data/processed_data/wbe_covid_n3_week.rds"))
############################################################################
# n = 2
wbe_covid_n2_week <- left_join(wbe_n2_week, covid_weekly, by = "week") %>% drop_na(week) %>% 
  mutate_at(vars(14:19),log10) %>% 
  mutate(A_N1_POS = n_pos_A_N1/n_A_N1,
         A_N2_POS = n_pos_A_N2/n_A_N2,
         B_N1_POS = n_pos_B_N1/n_B_N1,
         B_N2_POS = n_pos_B_N2/n_B_N2,
         C_N1_POS = n_pos_C_N1/n_C_N1,
         C_N2_POS = n_pos_C_N2/n_C_N2,
         log10_cases = log10(cases.reported)) %>% 
  mutate(date = as.Date(week)) %>% 
  mutate(qpcr = if_else(date > "2021-06-29","year2","year1"))

saveRDS(wbe_covid_n2_week,
        here("data/processed_data/wbe_covid_n2_week.rds"))
############################################################################
# n = 1
wbe_covid_n1_week <- left_join(wbe_n1_week, covid_weekly, by = "week") %>% drop_na(week) %>% 
  mutate_at(vars(14:19),log10) %>% 
  mutate(A_N1_POS = n_pos_A_N1/n_A_N1,
         A_N2_POS = n_pos_A_N2/n_A_N2,
         B_N1_POS = n_pos_B_N1/n_B_N1,
         B_N2_POS = n_pos_B_N2/n_B_N2,
         C_N1_POS = n_pos_C_N1/n_C_N1,
         C_N2_POS = n_pos_C_N2/n_C_N2,
         log10_cases = log10(cases.reported)) %>% 
  mutate(date = as.Date(week)) %>% 
  mutate(qpcr = if_else(date > "2021-06-29","year2","year1"))

saveRDS(wbe_covid_n1_week,
        here("data/processed_data/wbe_covid_n1_week.rds"))
```
